requirements.txt
# Core Framework
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
eventlet
python-dotenv==1.0.0

# Federated Learning & ML
flwr
tensorflow
numpy==2.4.0
pandas==2.3.3
scikit-learn

# Network Analysis & Security
nfstream
scapy
psutil

# Utilities
fpdf
websocket-client
